backend/requirements_new.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
mysqlclient==2.2.7
python-dotenv==1.1.1
django-cors-headers
requests>=2.31.0
google-generativeai>=0.3.0
Pillow>=9.0.0